+++ Thunderstore/CHANGELOG.md
@@ -1,5 +1,15 @@
 #### Changelog:
 
+`v0.5.1` -> `v0.5.3`
+
+```diff
++ ReplayV2: Much more accurate Replay Recorder and Replay Viewer
++ Replays now tries to submit 4 times if failed to submit
++ Added option to change the trombone pitch when playing at different game speed
++ Massively improved spectator performance
++ Several minor UI bug fixes for spectator mode
+```
+
 `v0.5.0` -> `v0.5.1`
 
 ```diff